+++ CHANGELOG.md
@@ -23,6 +23,12 @@
         * Two methods in `DataFetchersDelegateXxx`: one with the `DataLoader` parameter, that is used by the controller, and one without the `DataLoader` parameter, that is not used. The useless method is no more generated. This may result in compilation error, if the `@Override` method was added: you would then have to remove the implementation for this useless method.
 
 
+## Not released yet
+
+All modes (client and server):
+* The plugin now properly manages repeatable directives
+
+
 ## 2.9